CRUD ops & styling

+++ MyBooks/Controllers/BorrowingsController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using MyBooks.Data;
+using MyBooks.DTOs;
+using MyBooks.Helpers;
 using MyBooks.Models;
 
 namespace MyBooks.Controllers
@@ -47,66 +49,139 @@ namespace MyBooks.Controllers
         }
 
         // GET: Borrowings/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create(int? id)
         {
-            ViewData["BookID"] = new SelectList(_context.Books, "BookID", "BookID");
-            ViewData["UserID"] = new SelectList(_context.Users, "UserID", "UserID");
+            var books = await _context.Books.ToListAsync();
+            var users = await _context.Users.ToListAsync();
+
+            if (id != null)
+            {
+                ViewData["BookID"] = new SelectList(books, "BookID", "Book_name", id);
+            } 
+            else 
+            {
+                ViewData["BookID"] = new SelectList(books, "BookID", "Book_name");
+            }
+
+            User? LoggedInUser = SessionHelper.LoggedInUser(HttpContext);
+
+           
+            ViewData["UserID"] = LoggedInUser != null ? LoggedInUser.UserID : 0;
+            ViewData["Borrow_date"] = DateOnly.FromDateTime(DateTime.Today).ToString("yyyy-MM-dd");
             return View();
         }
 
         // POST: Borrowings/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("BorrowingID,BookStatus,Borrow_date,Return_date,Due_date,UserID,BookID")] Borrowing borrowing)
+        public async Task<IActionResult> Create([Bind("Borrow_date,UserID,BookID")] BorrowingDTO borrowingDto)
         {
-            if (ModelState.IsValid)
+            Borrowing borrowing = new Borrowing
             {
+                Borrow_date = borrowingDto.Borrow_date,
+                UserID = borrowingDto.UserID,
+                BookID = borrowingDto.BookID,
+                Due_date = borrowingDto.Borrow_date.AddDays(10)
+            };
+
+            if(borrowingDto.UserID == 0)
+            {
+                ModelState.AddModelError(string.Empty, "Du måste vara inloggad för att kunna låna böcker");
+                return View(borrowing);
+            }
+
+            if (!ModelState.IsValid)
+            {
+                ModelState.AddModelError(string.Empty, "Bokifno eller datum är felaktiga");
+                return View(borrowing);
+            }
+
+            var book = await _context.Books
+                .FirstOrDefaultAsync(m => m.BookID == borrowingDto.BookID);
+
+            
+
+            if (book == null || book.Available == false)
+            {
+                ModelState.AddModelError(string.Empty, "Boken är utlånad");
+                return View(borrowing);
+            }
+
+            if (borrowingDto.Borrow_date < DateOnly.FromDateTime(DateTime.Today))
+            {
+                ModelState.AddModelError("Borrow_date", "Lånedatum kan inte vara tidigare än idag");
+                return View(borrowing);
+            }
+
+            if (ModelState.IsValid) 
+            {
+                book.Available = false;
                 _context.Add(borrowing);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["BookID"] = new SelectList(_context.Books, "BookID", "BookID", borrowing.BookID);
-            ViewData["UserID"] = new SelectList(_context.Users, "UserID", "UserID", borrowing.UserID);
+
+            ViewData["BookID"] = new SelectList(_context.Books, "BookID", "Book_name", borrowing.BookID);
+            ViewData["UserID"] = new SelectList(_context.Users, "UserID", "User_Name", borrowing.UserID);
             return View(borrowing);
         }
 
+        // Get borrowed book
         // GET: Borrowings/Edit/5
-        public async Task<IActionResult> Edit(int? id)
+        public async Task<IActionResult> Borrow(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-
+            
             var borrowing = await _context.Borrowings.FindAsync(id);
             if (borrowing == null)
             {
                 return NotFound();
             }
-            ViewData["BookID"] = new SelectList(_context.Books, "BookID", "BookID", borrowing.BookID);
-            ViewData["UserID"] = new SelectList(_context.Users, "UserID", "UserID", borrowing.UserID);
+            ViewData["Return_date"] = DateOnly.FromDateTime(DateTime.Today).ToString("yyyy-MM-dd");
+            ViewData["BookID"] = new SelectList(_context.Books, "BookID", "Book_name", borrowing.BookID);
+            ViewData["UserID"] = new SelectList(_context.Users, "UserID", "User_Name", borrowing.UserID);
             return View(borrowing);
         }
 
+        // Return borrowed book
         // POST: Borrowings/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("BorrowingID,BookStatus,Borrow_date,Return_date,Due_date,UserID,BookID")] Borrowing borrowing)
+        public async Task<IActionResult> Borrow(int id, [Bind("BorrowingID,Return_date,BookID")] ReturnBookDTO borrowing)
         {
+           
             if (id != borrowing.BorrowingID)
             {
                 return NotFound();
+            }
+            var retunedBook = await _context.Borrowings.FindAsync(borrowing.BorrowingID);
+
+            if (retunedBook?.Return_date != null) 
+            {
+                ModelState.AddModelError(string.Empty, "Boken har redan återlämnats");
+                return View(retunedBook);
             }
 
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _context.Update(borrowing);
+                    var book = await _context.Books.FindAsync(borrowing.BookID);
+
+                    if (retunedBook is not null) 
+                    {
+                        retunedBook.Return_date = DateOnly.FromDateTime(DateTime.Today);
+                        _context.Update(retunedBook);
+                    }
+
+                    if (book is not null) 
+                    {
+                        book.Available = true;
+                    }
+                    
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -123,8 +198,7 @@ namespace MyBooks.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["BookID"] = new SelectList(_context.Books, "BookID", "BookID", borrowing.BookID);
-            ViewData["UserID"] = new SelectList(_context.Users, "UserID", "UserID", borrowing.UserID);
-            return View(borrowing);
+            return View(retunedBook);
         }
 
         // GET: Borrowings/Delete/5
@@ -153,6 +227,13 @@ namespace MyBooks.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var borrowing = await _context.Borrowings.FindAsync(id);
+
+            if (borrowing != null && borrowing.Return_date == null)
+            {
+                ModelState.AddModelError(string.Empty, "Boken måste återlämnas först innan lånedata kan raderas");
+                return View(borrowing);
+            }
+
             if (borrowing != null)
             {
                 _context.Borrowings.Remove(borrowing);
